Moved reservation acceptance to the admin main page. Also added functionality for reserving books as a user
