+++ src/App.jsx
@@ -4,7 +4,6 @@ import { useContext } from "react";
 import { Routes, Route, Navigate } from "react-router-dom";
 import Login from "./components/Login";
 import Main from "./components/Main";
-import Reservations from "./components/Reservations";
 import Header from "./components/Header";
 import AuthContext from "./store/auth-context";
 import Books from "./components/Books";
@@ -23,7 +22,6 @@ function App() {
         {isLoggedIn && <Route path="/register" element={<CreateAccount />} />}
         {isLoggedIn && <Route path="books" element={<Books />} />}
         {isLoggedIn && <Route path="employees" element={<Employees />} />}
-        {isLoggedIn && <Route path="reservations" element={<Reservations />} />}
         <Route
           path="*"
           element={<Navigate to={isLoggedIn ? "/main" : "/"} replace />}
