+++ src/index.css
@@ -91,6 +91,13 @@
     text-align: left;
 }
 
+.bookoverview-table thead tr.red {
+  background-color: #AB0303;
+  color: #ffffff;
+  text-align: left;
+}
+
+
 .bookoverview-table th,
 .bookoverview-table td {
     padding: 12px 15px;
@@ -108,6 +115,10 @@
     border-bottom: 2px solid #03AB03;
 }
 
+.bookoverview-table tbody tr:last-of-type.red {
+  border-bottom: 2px solid #AB0303;
+}
+
 .table-buttons {
     display: flex;
     justify-content: space-around;
@@ -121,6 +132,15 @@
 .icon:hover {
     cursor: pointer;
     box-shadow: 0 0 20px rgba(0, 0, 0, 0.15);
+}
+
+.icongreyedout {
+  font-size: 20px;
+}
+
+.icongreyedout:hover {
+  cursor: pointer;
+  box-shadow: 0 0 20px rgba(0, 0, 0, 0.15);
 }
 
 .basic-button {
