+++ src/styles/mainAdmin.css
@@ -1,48 +1,47 @@
-.adminheader-container{
-    Display: flex;
-  }
-  
-  .adminheader {
-    margin-left: 100px;
-    display: inline-block;
-  }
-  
-  .adminheader h4 {
-    color: #ab1403;
-    font-family: euclid-bold;
-    font-size: 20px;
-    margin-bottom: 10px;
-  }
-  
-  .adminheader h2 {
-      margin-top: 0;
-      font-family: euclid-bold;
-      font-size: 36px;
-    }
-  
-  .navbar {
-      margin-right: 100px;
-      margin-top: 25px;
-      text-align: right;
-      width: 100%;
-      font-size: 24px;
-    }
-    
-    .navbar span {
-      text-align:center;
-      margin-left: 10px;
-      margin-right: 50px;
-    }
-  
-    .navbar button {
-      height: 60px;
-      background: none;
-      border: none;
-      margin-left: 20px;
-    }
-    
-    .navbar button:hover {
-      padding: 10px 10px;
-      background-color: none;
-    }
-  
+.adminheader-container {
+  display: flex;
+}
+
+.adminheader {
+  margin-left: 100px;
+  display: inline-block;
+}
+
+.adminheader h4 {
+  color: #ab1403;
+  font-family: euclid-bold;
+  font-size: 20px;
+  margin-bottom: 10px;
+}
+
+.adminheader h2 {
+  margin-top: 0;
+  font-family: euclid-bold;
+  font-size: 36px;
+}
+
+.navbar {
+  margin-right: 100px;
+  margin-top: 25px;
+  text-align: right;
+  width: 100%;
+  font-size: 24px;
+}
+
+.navbar span {
+  text-align: center;
+  margin-left: 10px;
+  margin-right: 50px;
+}
+
+.navbar button {
+  height: 60px;
+  background: none;
+  border: none;
+  margin-left: 20px;
+}
+
+.navbar button:hover {
+  padding: 10px 10px;
+  background-color: none;
+}
